requirements.txt
ini-parser==1.2.1
praw==7.7.1